test: move rust module tests out of src
