+++ src/ui-host/BUILD.bazel
@@ -1,11 +1,41 @@
-load("//:defs.bzl", "yoyopod_rust_host_binary", "yoyopod_rust_host_test")
+load(
+    "//:defs.bzl",
+    "yoyopod_rust_host_binary",
+    "yoyopod_rust_host_integration_tests",
+    "yoyopod_rust_host_library",
+)
+
+yoyopod_rust_host_library(
+    name = "ui_host_lib",
+    srcs = glob(
+        ["src/**/*.rs"],
+        exclude = ["src/main.rs"],
+    ),
+    crate_name = "yoyopod_ui_host",
+)
 
 yoyopod_rust_host_binary(
     name = "yoyopod-ui-host",
-    srcs = glob(["src/**/*.rs"]),
+    srcs = ["src/main.rs"],
+    deps = [":ui_host_lib"],
 )
 
-yoyopod_rust_host_test(
+UI_HOST_TESTS = [
+    "framebuffer",
+    "hardware_whisplay",
+    "hub",
+    "input",
+    "lvgl_bridge",
+    "protocol",
+    "render_framebuffer",
+    "render_lvgl",
+    "runtime_state_machine",
+    "whisplay_panel",
+    "worker",
+]
+
+yoyopod_rust_host_integration_tests(
     name = "tests",
-    crate = ":yoyopod-ui-host",
+    tests = UI_HOST_TESTS,
+    deps = [":ui_host_lib"],
 )
