+++ src/ui-host/src/lib.rs
@@ -1,0 +1,11 @@
+pub mod framebuffer;
+pub mod hardware;
+pub mod hub;
+pub mod input;
+pub mod lvgl_bridge;
+pub mod protocol;
+pub mod render;
+pub mod runtime;
+pub mod screens;
+pub mod whisplay_panel;
+pub mod worker;
