+++ src/ui-host/src/render/framebuffer.rs
@@ -244,37 +244,3 @@ fn mix_rgb(foreground: (u8, u8, u8), background: (u8, u8, u8), weight: u8) -> (u
         (((foreground.2 as u16 * weight) + (background.2 as u16 * inverse)) / 255) as u8,
     )
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_scene_changes_with_counter() {
-        let mut first = Framebuffer::new(240, 280);
-        let mut second = Framebuffer::new(240, 280);
-
-        render_test_scene(&mut first, 1);
-        render_test_scene(&mut second, 2);
-
-        assert_ne!(first.as_be_bytes(), second.as_be_bytes());
-        assert_eq!(first.pixel(0, 0), rgb565(8, 10, 14));
-    }
-
-    #[test]
-    fn hub_fallback_uses_snapshot_accent() {
-        let mut first = Framebuffer::new(240, 280);
-        let mut second = Framebuffer::new(240, 280);
-        let mut first_snapshot = HubSnapshot::static_default();
-        let mut second_snapshot = HubSnapshot::static_default();
-        first_snapshot.accent = 0x00FF88;
-        second_snapshot.accent = 0x00D4FF;
-
-        render_hub_fallback(&mut first, &first_snapshot);
-        render_hub_fallback(&mut second, &second_snapshot);
-
-        assert_ne!(first.as_be_bytes(), second.as_be_bytes());
-        assert_eq!(first.pixel(72, 58), rgb565(0, 255, 136));
-        assert_eq!(second.pixel(72, 58), rgb565(0, 212, 255));
-    }
-}
