+++ src/ui-host/src/render/lvgl.rs
@@ -41,25 +41,3 @@ impl RendererState {
         self.active_screen
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn renderer_mode_names_are_stable() {
-        assert_eq!(RendererMode::Auto.as_str(), "auto");
-        assert_eq!(RendererMode::Lvgl.as_str(), "lvgl");
-        assert_eq!(RendererMode::Framebuffer.as_str(), "framebuffer");
-    }
-
-    #[test]
-    fn persistent_renderer_tracks_last_screen() {
-        let mut state = RendererState::default();
-
-        assert!(state.needs_rebuild(UiScreen::Hub));
-        state.mark_screen_built(UiScreen::Hub);
-        assert!(!state.needs_rebuild(UiScreen::Hub));
-        assert!(state.needs_rebuild(UiScreen::Listen));
-    }
-}
